move tx endpoint into wallet-specific code
